==> 00_startup.Rmd <==
---
title: "startup"
author: "Margaret Janiczek"
date: "2025-07-23"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Startup: One Time

## Request UKB RAP Access

## Set up project

## Wait for data to be dispensed


# Working on Data: To start

## Log in to UKB RAP

## Start Posit Workbench

Use HIGH PRIORITY (otherwise it will keep quitting and exiting and you lose all your updated code!!!)

## Clone GitHub repository of files 

In terminal:

git config --global user.name "YOURGITHUBUSERNAME"
git config --global user.email “YOURGITHUBEMAIL@email.com”

Then in console: 

`credentials::set_github_pat()`
And then paste in your github PAT token. 

Then back in terminal, clone your github repo with code for your project:
git clone https://github.com/mljaniczek/ukb_rap 

THEN make sure you open project (upper right corner), otherwise you won't be able to easily push your changes back to github! 

Now you're ready to actually code!!

## (Optional) Grab list of field names, modify them

This is useful if you want to just download a subset of the data to work with going forward. 

### List files in current project

We'll need to figure out where our files are in the RAP project storage.

```{r}
system("dx ls 'Showcase metadata'")
```

### Load the file using dxFUSE

The first way we can access a file in a project is by prepending a `/mnt/project/` to our file path. This lets us access files in our project. 

```{r}
field_data2 <- read.delim("/mnt/project/Showcase metadata/field.tsv")
```

```{r}
head(field_data2)
```

### Alternative: download file into Workspace Storage

If we want to be explicit in downloading files, we can utilize `dx download` within `system()` to download our files from the DX Toolkit.


```{r}
system("dx download '/Showcase metadata/field.tsv'")
list.files()
```

## Narrow down list of fields

To avoid downloading all 35K fields you probably want to narrow down your scope. Ahead of time for your analysis, identify which fields you need for your subset of data. 

You may also want to narrow down your cohort if you are not going to analyze all 500K UKB subjects, do that in Cohort Browser, separately. 

Below I have loaded in the fields available in my dataset, then filter down to make a few .txt files specific to the three separate data types that I use in my analysis. 

1. Phenotype data (age, BMI, smoking status, menopausal status etc)
- make pheno.txt, which should be a ONE COLUMN with the field names you want. Make sure to include the eid as well which will id the subject!! 

e.g. 

eid
p31
p48_i0
p49_i0

2. Metabolite data 


3. First Occurrences data



```{r}
field_dx <- read.csv("app81009_20240119215747.dataset.data_dictionary.csv")

library(tidyverse)

foccur <- c(130690:130835, 131270:131423) # endocrine, circulatory disorders

# additional vars to grab
additional <- c(23658 
                # the date of metabolite processing for p2 and p1 distinction
  
)



field_dx <- field_dx %>%
  mutate(field_mod = name) %>%
  mutate(field_mod2 = str_split_i(field_mod, "_", 1), 
         instance = str_split_i(field_mod, "_i", 2),
         field_mod3 = str_remove(field_mod2, "p"))

field_dx$instance[is.na(field_dx$instance)] <- 0 

field_dx_select <- field_dx %>%
  filter(str_detect(instance, "^0")) %>%
  filter(field_mod3 %in% 
           c("eid", as.numeric(char_key$field.showcase)
             #as.numeric(met_key$field.showcase),
            #as.numeric(med_key$field.showcase)
            )) %>%
 filter(entity == "participant")

# scratch why aren't blood pressure being selected

test <- char_key %>%
  mutate(field.showcase = as.numeric(field.showcase)) %>%
  filter(!(field.showcase %in% field_dx_select$field_mod3
           ))

test2 <- field_dx %>%
  filter(field_mod3 %in% 93)

# first occurrences data 
field_fo <- field_dx %>%
  filter(field_mod3 %in% c("eid", foccur))

field_fo2 <- field_fo %>%
  mutate(icd = str_to_lower(title)) %>%
  filter(entity == "participant") %>%
  select(name, icd, field_mod3) %>%
  mutate(
    cad = ifelse(str_detect(icd, c("i20|i21|i22|i23|i24|i25|z951|z955")),1,0),
    afib = ifelse(str_detect(icd, c("i48")), 1, 0),
    stroke = ifelse(str_detect(icd, c("i60|i61|i629|i63|i64|i678|i690|i693|g951|h341|h342|s066")),1,0),
    hypertension = ifelse(
      str_detect(
        icd, c("i10|i11|i12|i13|i15|o10")
      ), 1,0
    ),
    diabetes_other = ifelse(str_detect(
      icd, c("e10|e12|e13|e14") # separated out other diabetes 11/22/23
    ), 1, 0),
    diabetes_t2 = ifelse(str_detect(
      icd, c("e11")
    ), 1, 0),
    any = cad + afib + stroke + hypertension + diabetes_other + diabetes_t2
  )

field_fo3 <- field_fo2 %>%
  filter(any == 1)

write.csv(field_fo3, file = "first_occ_key_exporter.csv")

# testing proteomics for ben 
# 
# field_dx_select2 <- field_dx %>%
#   filter(entity == "olink_instance_0")
# 
# write.csv(field_dx_select, file = "please_work.csv")
# 
# write.csv(field_dx_select2, file = "please_work_for_ben.csv")
```


### Load the file into R

```{r}
field_data <- read.delim("field.tsv")
```

```{r}
head(field_data)
```

## Load in the keys for fields used in analysis on Nightengale

```{r}
library(dplyr)
#load("ukb_keys.Rdata") for all pheno data

load("analysis_variable_keys.Rdata")

field_data_select <- field_data2 %>%
  filter(field_id %in% c(as.numeric(char_key$field.showcase), as.numeric(met_key$field.showcase), as.numeric(med_key$field.showcase)))


length(unique(field_data_select$field_id))
  
```


### Writing our file to the RStudio Workspace

```{r}
write.csv(field_data_select, file="updated_field_data.csv")
list.files()


```

### Saving Our file back into Project Storage

```{r}
system("dx upload updated_field_data.csv")
```


## Use Table Exporter to save a .csv file of the data to your Project space


## Wrap up

### Commit and push any file changes


### Save results back to project folder 


### Terminate RStudio session

Be sure to terminate the Posit workbench or it will just keep running (and costing money!!)

# Working on data: subsequent

## Log in to UKB RAP

## (Optional) Use GitHub repository

## (Optional) Connect to GitHub so you can push file changes

## Load in .csv file of data with selected fields

## Run analysis

## Wrap up

### Commit and push any file changes


### Save results back to project folder 


### Terminate RStudio session

Be sure to terminate the Posit workbench or it will just keep running (and costing money!!)


## Scratch

```{r}
pheno_text <- field_dx_select %>%
  mutate(text_column = str_replace_all(title, "[^[:alnum:]]", ".")) %>%
  left_join(char_key, by = c("field_mod3" = "field.showcase")) %>%
  select(name, title, text_column, field_mod3, col.name)
```
