==> 01_data_load.Rmd <==
---
title: "Untitled"
output: html_document
date: "2025-08-21"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

install.packages("readr")
install.packages("gtsummary")
install.packages("glue")
library(gtsummary)
library(dplyr)
library(stringr)
library(tidyr)
library(readr)
```

```{r}

# the following three files were prepared using table exporter and the .txt files
#"pheno.txt"
#"mets.txt"
#"first_occurrences.txt"

system('dx download pheno2.csv') # downloaded phenotype data
system('dx download mets.csv')
system('dx download first_occurrences.csv') 

pheno <- read_csv("pheno.csv", name_repair = "universal_quiet")

mets <- read_csv("mets.csv", name_repair = "universal_quiet")

# repair names
fo <- read_csv("first_occurrences.csv", name_repair = "universal_quiet") 






# grab columns of occurrence
# make column keys based on source vs date
event_key <- colnames(fo)[str_detect(colnames(fo), "Source")]
eventdate_key <- colnames(fo%>% select(-c( Sample.Measured.Date.and.Time...Instance.0)))[str_detect(colnames(fo%>% select(-c(Sample.Measured.Date.and.Time...Instance.0))), "Date")]

event <- fo %>%
  select(Participant.ID, all_of(event_key))

event_date <- fo  %>% select(Participant.ID, all_of(eventdate_key))

# go to longer
event_long <- event %>%
  pivot_longer(!Participant.ID, values_drop_na = TRUE) %>%
  group_by(Participant.ID) %>%
  mutate(id = row_number())

# grab date of first occurence
eventdate_long <- event_date %>%
  pivot_longer(!c(Participant.ID, Date.of.attending.assessment.centre...Instance.0), values_drop_na = TRUE) %>%
  group_by(Participant.ID) %>%
  mutate(id = row_number()) %>%
  mutate(afterassessment_fo = value > Date.of.attending.assessment.centre...Instance.0,
         beforeassessment_fo = value < Date.of.attending.assessment.centre...Instance.0)

# now grab icd10 codes

event_long <- event_long %>%
  mutate(name = str_to_lower(name)) %>%
  mutate(
    cad = ifelse(str_detect(name, c("i20|i21|i22|i23|i24|i25|z951|z955")),1,0),
    afib = ifelse(str_detect(name, c("i48")), 1, 0),
    stroke = ifelse(str_detect(name, c("i60|i61|i629|i63|i64|i678|i690|i693|g951|h341|h342|s066")),1,0),
    hypertension = ifelse(
      str_detect(
        name, c("i10|i11|i12|i13|i15|o10")
      ), 1,0
    ),
    diabetes_other = ifelse(str_detect(
      name, c("e10|e12|e13|e14") # separated out other diabetes 11/22/23
    ), 1, 0),
    diabetes_t2 = ifelse(str_detect(
      name, c("e11")
    ), 1, 0)
  )

# now have one var for disease
test <- event_long %>%
  pivot_longer(!(Participant.ID:id), names_to = "disease_fo", 
values_to = "has_disease_fo",
values_drop_na = FALSE) %>%
  group_by(Participant.ID, disease_fo, has_disease_fo) %>%
  slice(1) # just grab first instance of disease

test3 <- test %>%
  select(Participant.ID, id, source_of_report = value, disease_fo, has_disease_fo) %>%
  filter(has_disease_fo == 1) %>%
  group_by(Participant.ID, disease_fo) %>%
  slice(1)

final_fo <- eventdate_long %>%
  right_join(test3)

prev <- final_fo %>%
  filter(beforeassessment_fo == TRUE) %>%
  select(Participant.ID, disease_fo, has_disease_fo) %>%
  pivot_wider(names_from = disease_fo, values_from = has_disease_fo,
              values_fill = 0, names_prefix = "prevalent_")

inc <- final_fo %>%
  filter(afterassessment_fo == TRUE) %>%
  select(Participant.ID, disease_fo, has_disease_fo) %>%
  pivot_wider(names_from = disease_fo, values_from = has_disease_fo,
              values_fill = 0, names_prefix = "incident_")


final_fo2 <- prev %>%
  full_join(inc) %>%
  full_join(fo %>% select(Participant.ID)) %>%
  replace(is.na(.), 0)


write.csv(final_fo2, file="first_occurrences_processed.csv")
system("dx upload first_occurrences_processed.csv")
# 
# library(ggplot2)
# final_fo %>%
#   filter(afterassessment_fo == TRUE) %>%
#   filter(has_disease_fo == 1) %>%
#   ggplot(aes(disease_fo)) +
#   geom_bar(aes(fill = source_of_report)) +
#   theme_bw() +
#   ylab("Percent") +
#   xlab("Disease")


final_fo %>%
  filter(has_disease_fo == 1) %>%
  ungroup() %>%
  select(disease_fo, afterassessment_fo) %>%
  tbl_summary(by = afterassessment_fo)
```
